ONA: Refactor notification builder and register broadcast receiver.
Refactors the posting of the ONA notification by using a notification
builder from the framework facade. The notification building is isolated
in a helper class. OpenNetworkNotifier registers a broadcast receiver to
listen to user actions like swiping to dismiss and notification content
clicks. Also plumbs in a clock to stop using System#currentTimeMillis.

Bug: 38460614
Bug: 37357441
Bug: 62410249
Bug: 63544378
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: Ifa1eef1f17c3f3271ec4a11f4311ea00a2698be3