WifiStateMachine: disable attempt to id gateway
When transitioning to ConnectedState from L2ConnectedState, we attempt
to identify the gateway.  This is intended to be used for linking SSIDs
(overly simplified: different SSIDs that are from the same AP).

This call is doomed to fail and has not been working since
multi-networking support was added.  The consequence of this call is
that we can process messages related to mode switches out of order and
end up "stuck" on or off.

This call is commented out in this CL and further cleanup will be done
with the WifiStateMachine refactor.  A bug was filed to track work to
implement this functionality with information from ConnectivityService.

Bug: 64118984
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: integration tests

Change-Id: I65ea02cd9b53b2e563d221e59581c7c11a9c6a43