[AWARE] Defer requests received while terminating an NDP
Defer requests received while an identical (canonically)
requests is being terminated (i.e. a termination was issued
but confirmation of termination not yet received).

In such a case the request is now rejected but a flag is set which
when (any) termination confirmation is received tickles the
connectivity service to re-evaluate the requests.

Bug: 63866251
Test: unit tests and integration tests
Change-Id: I77f8fd2c8444fb50aa05c356645216622f282677