WifiMetrics: add methods to increment ONA metrics.
Tracks the number of notifications posted for each state of ONA, and the
number of user actions made for each notification type. Also tracks the
state of the feature setting, the size of the recommender blacklist,
recommendation update counts, and connection initiation failure counts.

Also adds additional debug logging.

Bug: 37501326
Bug: 66996715
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
and verify in dumpsys

Change-Id: Ia00ae95fbcf923ebae051ba82cd3174cfe11be31