Docs: Adds brief instructions for how to do Wi-Fi scan. Bug: 67967771 Test: Ran make ds-docs.
We don't actually document how to do a Wi-Fi scan anywhere on DAC;
at the moment, this seems like the place where it makes the most sense
to provide that guidance.

The forcing function for doing this is that a new doc explicitly
recommends that devs do Wi-Fi scans in certain cases, but doesn't
have a place to link to that tells them how to do so. That doc
plans to send readers here.

Change-Id: Ic35fb77d323e825e5bd5b4eec2716730f0d531e1