WifiNative: New interface setup/teardown stub methods
Adding stub methods for the new interface setup/teardown & status change
callbacks.

a) initialize() - Used to initialize and register for death
notifications from HAL & wificond.
b) setupInterfaceForClient/SoftApMode() - Used to configure an interface in
the specified mode. This method accepts a listener object which will be
invoked when the interface is up, down or destroyed. Based on the device
configuration (Iface combo supported) & HalDeviceManager policy, any
existing interface might be torn down. The method will return an interface
name on success which can be used as handle for further operations on
that interface.
c) teardownInterface() - Used to teardown an interface when the mode is
no longer needed (For ex: when the user turns off wifi/SoftAp). This
will also trigger the associated |destroyedListener|.
d) registerStatusListener - Used to listen for any change of status in
the native daemons. (For ex: wificond crashing).

Bug: 69265194
Test: Compiles
Change-Id: I405be2e5624444471faa09a358312ee02b5ee73b