WifiVendorHal: Support concurrent interfaces
Changes in the CL:
a) Add a map of iface name to corresponding handles for all STa & AP
ifaces.
b) Add iface name as input argument to all the HAL methods operating on an
iface object.

Modified unit tests and WifiNative to use the default interface name for
all operations.

Bug: 69270375
Test: Unit tests
Test: Device boots up and connects to wifi networks.
Change-Id: If174cb25c5415b0b2633454a5cc5378821c837fa