Release the correct WiFi Multicast Wakelock
In current implementation, when the same UID allocates, and acquires
multiple WiFi Multicast wakelocks the tag associated with the lock is
passed in the acquire method, and stored in the multicasters list.
However, when the UID tries to release the lock, the tag is not included
in the method call, and hence all the wakelocks found in the list of
multicasters for this UID are released.
This behavior does not reflect what the user app expects since only one
multicast wakelock should be released and the others should still be  acquired.

This commit fixes this situation by releasing only the proper wakelock.

Bug: 70691435
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I25c5ec0e03cd2a0572f4275005b205900a649ba7