Add onboarding process for Wifi Wake.
Creates a notification to display to the user the first time they turn
off wifi with the feature enabled. Updates the ConfigStoreData to
include a bit for "isOnboarded". Creates a factory class to handle
constructing the actual notification. Gates handleScanResults on user
onboarded status.

Bug: 64094365
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: adb push old WifiConfigStore.xml; adb shell stop; adb shell start;
verify that the config store loads properly and writes the new
WakeupConfigStoreData

Change-Id: Ia3949a70a2c64bc40fc658788e395061a7ca86ee