WifiStateMachinePrime: activate scan mode
Activate control of ScanOnlyMode via WifiStateMachinePrime.  Note: this
still has ScanMode in WifiStateMachine as a shadow state, control of
interfaces and other state is managed by WifiStateMachinePrime and
ScanOnlyModeManager itself.

This CL additionally adds the dependence of supplicant to scan mode.
This is required to properly set up the new interface.

This CL also incorporates changes to handling Messages for blocking API
calls in the WpsRunningState.  These are no longer dropped during WPS
and are now deferred (as documented in the existing comment).

NOTE: A follow-on CL will move app scan request handling from WSM to the
mode manager.  An additional CL will move start for WifiWake to
ScanModeManager.

Bug: 32661064
Test: manual testing via toggles in settings
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: locally ran new scanning integration tests
Test: wifi integration tests
Change-Id: Ie0cfb856cf3ad20b1d3ffb260719de5108afa730