ScanOnlyModeManager: Move to new interface mgmt methods
Migration of ScanOnlyModeManager to use the new interface
management methods in WifiNative.

Changes in the CL:
a) Use WifiNative.setupInterface() & WifiNative.teardownInterface()
to setup an interface in the specific mode.  This handles starting
HAL if needed. Also, sets up the interface in wificond, HAL.
b) Removed the hardcoding of interface names in ScanOnlyModeManager.
c) Removed wificond/HAL death handling and instead will use the new
WifiNative.StatusListener to listen for any daemon's death.

Bug: 69426063
Test: Unit tests.
Change-Id: I2c9acab94c90e6d03c2c79437868f99149656520