WifiStateMachinePrime: collapse modes to a single state
With the ModeManagers becoming fully disposable, we no longer need the
idle states.  Collapse the idle + active states for each mode in to a
single state.  Also adds a DefaultModeManager that will handle calls
servicing WifiManager API calls.

In the case of a failure (scan/softap modes currently), we should move
to the disabled state.  Calls will be handled by the DefaultModeManager
when wifi is fully off.

Bug: 71784953
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: wifi integration tests
Change-Id: I7fdd8994726616e8b51f3fae6887c73da5da883a