WifiServiceImpl: directly call WifiApConfigStore
Directly call WifiApConfigStore (on the WSM thread) when
get/setWifiApConfiguration is called.  We do not require that this be
plumbed through SoftApManager since ap config retrieval is decoupled
from softap startup.  Tethering is explicitly stopped and restarted
when the config is changed in Settings.

This additionally addresses a race with setting/getting the ap config.

Bug: 71784834
Bug: 72070781
Test: manually verified configs viewed and modified in settings
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: wifi integration tests
Change-Id: I176803f962aa671574e9c110b83594a39a578716