Mac Randomization: Wifi Framework
Implementation of Connected MAC Randomization on the Wifi framework
side. The state machine listens for global setting
WIFI_CONNECTED_MAC_RANDOMIZATION_ENABLED to decide whether to randomize
MAC addresses or not.

After deciding to randomize, the state machine sets MAC address through
WifiNative, WifiCondControl, and then WifiCond. Then, we bring the
interface down, change MAC address, and bring the interface back up.

In WifiNative/ConnectivityManager, we will need to suppress the
interface call back for going down/up, so that the rest of the framework
does not react to the status change (b/72459123).

Bug: 63905794
Bug: 71548300
Test: Manual Testing. Will add unittests later
Change-Id: I87aa48e580cb13717ef487559b7b33b65db841e6