SupplicantP2pIfaceHal: Move to an on-demand iface creation model
Currently, we setup the p2p iface whenever we establish the HIDL
connection to wpa_supplicant. As a part of moving to an on-demand
iface creation model, don't create the interface immediately. Will
expose a new API to be used by WifiP2pNative to actually create the
interface.

Changes in the CL:
a) Don't initialize P2p iface in wpa_supplicant as soon as
"wpa_supplicant" HIDL service is available.
b) The new SupplicantStaIfaceHal.setupIface() & teardownIface() methods
will be invoked by WifiP2pNative when we need to setup the interface.

Note: These changes in SupplicantP2pIfaceHal are similar to the changes
made in SupplicantStaIfaceHal for supporting concurrency.

Bug: 72497243
Test: Reworked existing |setup| unit tests & added a new test for |teardown|.
Test: Ensured that p2p works using settings after toggling wifi.
Change-Id: I7092c787a438e10d6d94032c93340c20aaeafa1e