WifiNative: Refactor teardownAllInterfaces()
There is already a version of this functionality in onNativeDaemonDeath.
Reuse the functionality from there. This also prevents
the concurrent modification exception in the current implementation
(because we were not using the ifaceIdIter to remove from the map).

Note: We don't need to explicitly call mWifiVendorHal.removeIface()
because mWifiVendorHal.stopVendorHal() takes care of clearing
everything.

Bug: 73080100
Test: Unit tests
Change-Id: I68caf1d64a28e137efda0a6c73df4eb49402f074