WifiStateMachinePrime: add tests and remove individual iface teardown
Before fully switching to dynamic iface management, we must first rely
on the hammer of removing all interfaces when entering a new mode.  This
prevents exit from a mode from tearing down the next mode due to a
single iface name.

This CL also adds additional unit tests to boost coverage.

Note: this CL also removes the check for teardownAllInterfaces from WSMP
startup since this is now covered in lower level creation.  We will call
this when entering disabled state, but not in the WSMP constructor.

Bug: 72996543
Test: frameworks/opt/net/wifi/tests/wifitests/coverage.sh
Test: verified wifi wake enables
Change-Id: I8abff0d158809fbeba0a35326c84aadb0f85644b