LogcatLog: add tracing functionality to trace()
Some people have the crazy idea that the trace()
method should help trace program execution.

I don't know why they would think that, but
I'll relent and make it so.

Along the way: refactor how we filter trace/debug
messages, so that we don't pay the cost of
a stack trace unnecessarily.

Bug: 73972044
Bug: 30737821
Test: tests/wifitests/runtests.sh (on walleye)

Change-Id: Ida435642af81945bfcb16db5cf39b20e5bebc5bb