WifiVendorHal: fix use of mLog.err()
In order to minimize the cost of disabled log
messages, the WifiLog API asks its caller to
pass in parameters using individual method calls,
rather than stringifying the whole logging message.

Fix one case in WifiVendorHal, where the caller
was stringifying the log message. Also, modify one
test, which was not specifying the format string
for the log message.

Bug: 73971539
Test: tests/wifitests/runtests.sh (on walleye)
Change-Id: I91fdcc6b447fb96ba9423680e833e6f85544fc06