WifiP2pNative: Handle stale interface destroyed invocation
When the P2P interface is destroyed by the WifiP2pService, we don't need
to process the associated interface destroyed invocation. This listener will
otherwise trigger a second cleanup. We need to handle the destroyed
callback only if the HalDeviceManager removed the interface on it's own (
i.e if AP or Aware interface was brought up & we need to destroy P2P interface).
Also, added some extra logs to help in debugging issues in this area.

The problem with links camera OTA update:
a) links camera closes the P2P channel -> WifiP2pManager.close()
b) This triggers WifiP2pNative.teardownInterface()
c) links camera then reopens the P2P channel ->
WifiP2pManager.initialize()
d) This triggers WifiP2pNative.setupInterface()
e) The |InterfaceDestroyedListener.onDestroyed| callback for the teardown
initiated in (b) comes in at this stage and triggers a cleanup of the iface
setup in (d).
f) The |InterfaceAvailableForRequestListener.onAvailabilityChanged|
callback then sees IWifiP2pIface == null and declares that the interface
is not available.

With this fix, step (e) will now become a no-op.

Bug: 73787760
Test: Ran CtsVerifier tests
Test: Verified that links camera's P2P connection still works.
Test: Need to verify if this fixes the OTA issue seen by links camera
(Not able to repro locally)

Change-Id: I3db40baa898234e9bc4297e57c18f839c52e9f4e