SupplicantStaIfaceHal: Register death notification only on ISupplicant
Currently, we register for death notifications on both ISupplicant & the
ISupplicantStaIfaceHal remote objects. Since these are hosted on the
same process (wpa_supplicant), we should register for death
notifications only for the top level object.

PS: Intentionally or not, this is what we already do in WificondControl
& HalDeviceManager. We don't register for death notifications for the
child objects.

Bug: 75993335
Test: Unit tests
Test: Killed wpa_supplicant and ensured that only one death notification
is sent to WifiNative.

Change-Id: I1a6610e45c6b478d17cfff2b5ee5edd2efc33a5c