WifiStateMachine: Act on reported EapFailure error
This commit acts on the EAP Failure error code recieved from
wpa_supplicant. The receipt of AUTHENTICATION_FAILURE_EVENT with reason
ERROR_AUTH_FAILIRE_EAP_FAILURE will trigger connect mode to check the
error code, and if it matches the vendor specific code for certificate
expiry, it will call into TelephonyManager to trigger the download of a
new certificate.

Bug: 64612561
Test: Unit Test
Test: Regression Test

Change-Id: I52b0a53504f2e50e91c0f19d987736d08699550c
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>