ActiveModeManager: add default scan available bcast
Add a default implementation to send the ScanAvailable
broadcast for mode implementations to use - in particular, when they are
entering a mode.  This allows modes to signal the proper state to
WifiScanningService without adding ref counting that would complicate
the ScanningService impl when only one STA interface is available.  At
the point multiple STA interfaces can be running in parallel, this will
need to be changed.

Note: the scan available update in WSMP for softap mode will be removed
when softap is independent of the other modes.

Bug: 76168955
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manually verified WifiScanner state when toggling wifi, location
and softap
Test: wifi integration tests

Change-Id: Ie6857478aa1082e3b9886a2497916225a56296c2