Wifi Usability Metrics
Collect metrics regarding Tx/Rx counts to see when there is a data
stall. We'll be collecting Tx successes, Tx retries, Tx failures, and Rx
successes, and we'll generate WifiIsUnusableEvent when we detect a data
stall.

Data stall events are triggered when 1) there is an increase in Tx
failures or 2) there is a significant increase in Tx success, but Rx
success has not changed. To prevent data stall events from flooding the
metrics, there is a time-based throttle and a limit on the total number
of events.

Also, generate WifiIsUnusableEvent whenever there is a firmware alert.

This metric can be enabled by changing WIFI_IS_UNUSABLE_EVENT_METRICS_ENABLED
settings value to 1. By default, WifiIsUnusableEvent is not collected.

Manual Check Process:
- Connect to a network
- adb shell dumpsys wifi | grep -10 WifiIsUnusableEventList
- adb shell settings put global wifi_is_unusable_event_metrics_enabled 1
- adb shell dumpsys wifi wifiMetricsProto
- Move away from the ap to trigger data stall
- adb shell dumpsys wifi | grep -10 WifiIsUnusableEventList
- adb shell dumpsys wifi | grep -10 mExperimentValues
Bug: 77603216
Test: Unittest, manual check
Change-Id: Ie926776ddd8e8e81142bcd2e8c37d46b7bf31590