WifiConfigManager: Temporarily disable network with no internet access
Changes in the CL:
a) Add a new network disable reason to temporarily blacklist an
auto-connected network.
b) When the device auto-connects to a network and then it detects a no
internet report, we temporarily disable the network for 10 minutes. This
should force network selection to pick another candidate if available.

Note:
a) If there are no other candidates available to auto-connect, the device
will remain connected to the existing network (even though it is
temporarily disabled).
b) If the user explicitly clicked on the network, we'll not temporarily
disable the network.

Bug: 72635747
Test: Unit tests
Test: Manually verified that device switched away from an
auto-connected network when it detects no internet.
Change-Id: I644693fa5d4477bdce1d39f391fc2fb24a773eeb