Update WifiInfo MAC address more frequently
Currently, we reset MAC address stored in WifiInfo to 02:00:00:00:00:00
whenever we are handling disconnect with Connected MAC Randomization
enabled. The disconnect handling can happen after we have started
connecting to a new network and generated a new randomized MAC address.

Update the MAC address to the real value once we have successfully
connected to the new network.

Bug: 78013990
Test: Manual Check, Unittest
Change-Id: I8f3d66bef353e91d3938a74c5d0f10d746903961