WifiStateMachine: send supplicant connection bcast
When tearing down client mode, add broadcast indicating that the
supplicant connection state has changed.  While this is often misused to
indicate wifi connection state, we cannot remove it yet.  The broadcast
has been deprecated and will be removed in the future, but for now, we
still send the broadcast at the start and end of client mode.

Bug: 79477256
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: Id5dc4f4debf653968d8dca702c42fef89a5ad642