[RTT] Recreate RTT controller when HAL indicates it is invalid
When there is a HAL mode transition the RTT controller is destroyed.
Mode transition is STA - SAP and back transitions for non-DBS devices.

Modification to allow framework to detect when the controller is
currently invalid and try to recreate it. Will allow recovery for
non-DBS devices: otherwise RTT would only recover when Wi-Fi is
toggled.

Note: not a full solution since the availability APIs (broadcast +
API) do not perform correctly in such transitions. That fix will
be deeper and require HalDeviceManager mods.

Bug: 111218083
Test: (new) unit test
Test: new ACTS - RangeApSupporting11McTest:test_rtt_in_and_after_softap_mode
Change-Id: I6d28ef7a50b02fe42e23b9172621bd83452658f1