Always generate wifi metrics incrementally
When generating metrics, we always want to be incremental, meaning that
current metrics should be cleared after they are uploaded. This ensures
that data is only logged once.

Also, stop using automaticBugReportTaken field in ConnectionEvent as any
ConnectionEvent that got uploaded will be cleared in clear() method.
This proto field will be removed in a separate bug b/111395361.

Bug: 79580512
Test: Unittest
Test: Manual check that ConnectionEvents are correctly cleared
Test: Connect to at least 2 networks
Test: adb shell dumpsys wifi | grep -A5 mConnectionEvents
Test: adb shell dumpsys wifi wifiMetricsProto
Test: Connect to 1 network
Test: adb shell dumpsys wifi | grep -A5 mConnectionEvents
Test: Verify that only 1 ConnectionEvent is logged
Change-Id: Iad464e10b8190e740464131db54164767085e9a5