WifiController: Enable softap toggle in airplane mode
Wifi tethering should work even in airplane mode. So, remove
the restriction in WifiController which prevents enabling SoftAp in
airplane mode. There is no FCC or carrier requirements to disable SoftAp
in airplane mode. LTE tethering would obviously disabled because the
mobile radio is off in airplane mode.

Note: We need some UI changes for this to be available to the users
(details mentioned in the associated bug).

Bug: 111681176
Test: Unit tests
Test: Added a new ACTS test which toggle softap state after turning on
airplane mode.

Change-Id: Ie67b1de1d31615a75683c1e91290e647eaae809c