WifiNative: Enable Wifimonitor in scan mode
Even though WifiMonitor was generally designed for handling supplicant
(which is not running in scan mode now) events, there are a couple of
wificond related events using WifiMonitor. Wificond events include scan
completed event. So, even though the scan was completed by wificond, the
events were not getting delivered to WifiScanningService.

Note: This was a regression introduced by ag/4376305. Not sure why this wasn't
caught in the integration tests for that CL.

Bug: 112442032
Test: Unit tests
Test: testWifiManagerScanWhenWifiOffLocationTurnedOn CTS test passes
now.

Change-Id: I961703e54322bbd9748d2816c1c781e3de715c0d