ExternalCallbackTracker: Unlink binder token death notification
Ensure that we unlink the binder token associated when we replace the
callback object.
Currently settings uses the same callback object (so same
callbackIdentifier) for all their SoftAp callback registrations. This
would result in a whole lot of binder tokens which can't be GC'ed
because of the death notification linking.

Bug: 110347655
Test: Unit tests
Test: Manual verification by adding logs in reset method.
Test: Ensured that softap & traffic poller works.
Change-Id: I880700358633b54e6b76a473adb84d6fa8c0d5c7