Resetting SIM identity after store is read
Sometimes CMD_RESET_SIM_NETWORKS is received before CMD_BOOT_COMPLETED.
And then identity is not reset because stored data is not loaded yet in
this timing. Hence, framework requests connection with old identity.

This patch calls the method resetSimNetworksresets after store is read
to make sure identity is updated regardless of the order of events.

Bug: 113547031
Test: Manual
Test: run unit tests
Change-Id: I601bc725ab102867dbb5441617d3f4d76321a95f