WifiNetworkFactory: Always allow requests from signature app
For apps like settings/sl4a (needed for ACTS tests) which run with
SYSTEM_UID, we cannot determine the importance (FG app/service, etc)
using the provided API's. Any app which holds the NETWORK_SETTINGS
permission should be able to bypass all the foreground checks.

This is probably going to be only used for ACTS tests, since we don't
expect settings to use this API surface for now.

Bug: 113878056
Test: Unit tests
Change-Id: Ie1a27cd5b365a206c4d81d38837d026cac060892