Add functionality to hi-perf WifiLocks
In current implementation, apps can create and acquire WifiLocks
of different types. However, there is no functionality associated
with these locks.
This commit provides this missing functionality for hi-perf locks
by disabling power saving mechanism upon acquisition, and enabling
it back upon release.
This behavior happens for both screen-on and screen-off, and no
matter if app is running in foreground or background. This is consistent
with the API documentation for hi-perf WifiLocks.

Bug: 34905427
Test: Unit test
Test: ./wifi/tests/wifitests/runtests.sh

Change-Id: I5d68c0376b69f9726b83fd56469e5d397845dd8d