Pno recency sorting
WifiConfigManager: Modify PNO list ordering

The PNO network list size that most drivers support is limited to 16, so
we sort the user's saved networks to ensure that the most useful ones
for the user is in the PNO watch list.
The current logic sorts the networks based on the number of previous
connection attempts (any ties are resolved based on when it was last seen in
scan results).
This has now been modified to move the last connected network to the
top of the list followed by the previous sorting algo. This will ensure
that we always auto-connect back immediately (when screen off) if there
is any intermittent disconnection.

Bug: 75956033
Test: compile, unit tests
Change-Id: I5a3fcb232d48b6f5d7db4929c9e0eca4de63e017