Network-Suggestion/Request: Handle saved network with same credentials
For NetworkRequest processing, CONNECT_NETWORK message already handles
connecting to an existing saved network (using networkId). So, just use
the network ID of the existing saved network to request connection. If
it doesn't exist add it to WifiConfigManager & then use the allocated
network ID to request connection.

For NetworkSuggestion processing, if the network already exists as a
saved network, skip attempting to add the network configuration in the
evaluator.

Bug: 120581555
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: act.py -c wifi_manager.config -tb dut-name -tc WifiNetworkSuggestionTest
Test: act.py -c wifi_manager.config -tb dut-name -tc WifiNetworkRequestTest
Change-Id: If7e517689c4edd72a30eab8e14ee52010a04ca33