WifiConfigManager: Permanently disable network with no subscription
This CL disables EAP network if EAP authentication fails with
AT_NOTIFICATION notification code (1031) which indicates user has not
subscribed to the requested service. This typically happens if user
has not subscribed carrier wifi service but wifi configuration to the
serivce is pre-configured. In this case, the device was trying to
connect to such EAP network periodically even if connection was
always unsuccessful.

Bug: 122921493
Test: Unit tests
Test: Manually verified that device will not try to connect to the
EAP network if EAP authentication failed with 1031 code before.
Signed-off-by: Daichi Ueura <daichi.ueura@sony.com>
Change-Id: I66e761f5b778e3d9f5aceac9ae0f1e9997d59c53