WifiNetworkSuggestionsManager: Post notification for user approval
Whenever a network suggested by an app is available in scan results,
post a notification asking to approve the app (if not already approved).
If there are multiple such apps matching a suggestion & pending approval,
then we pick the first one to post the notification for. Once a notification is posted,
we don't post any further notification until the user interacts with the
active notification (i.e dismiss, allow, deny).

a) If the user dismisses the notification, we popup the notification for
the next time a suggestion from the app is available in scan results.
b) If the user clicks on allow, then we don't popup notifications for
any further connection from the app.
c) If user clicks on disallow, then we revoke CHANGE_WIFI_STATE app-ops,
which will trigger a full removal of the app & it's suggestions from the
database.

Bug: 120072329
Test: Manually verified the notification posting using ACTS tests.
Test: Manual verification of interacting with the notification.
Test: `act.py -c wifi_manager.config -tb dut-name -tc
WifiNetworkSuggestionTest`

Change-Id: I4a3ed6abca01aab84fcd3717fde56bca6187fae6