passpoint: fix the matching logic for EAP-method in NAI realm
If we can't find matched auth method from authParams with
Passpoint credential such as (EAP-TTLS and AuthMethod:PAP),
we consider the AP is not matched with the profile.
In the case that auth params is empty in NAI realm element,
we can't find any matched auth method from the collection.
So, It eventually causes the passpoint profile not to match with this AP
even if the domain name is already matched.

Bug: 122666189
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: tested it with Passpoint AP that has EAP method and no auth
params.
Signed-off-by: Ecco Park <eccopark@google.com>

Change-Id: Iec90ffc64af0cc66870ab290cd374148135cb8e5