ClientModeImpl: Remove internet capability for network request
When we connect to a network in response to a network request, ensure we
remove the INTERNET capability from the agent created. This is done to
limit the network request API surface to local connections only (i.e the
default network on the device will not switch to wifi if there are other
agents available).

Also,
a) Moved all network capabilities setting to a common method in
ClientModeImpl.
b) Removed unnecessary variables holding network capabilities in
ClientModeImpl.

Bug: 119632465
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: act.py -c wifi_manager.config -tb dut-name -tc WifiNetworkRequestTest
Test: New CtsVerifier tests for network request.
Test: Will send for full regression testing.
Change-Id: Idc133354ecc31ff6dfd38800763ae9d935655476