HalDeviceManager: Track isReady() correctly
With ag/6150830 in, mWifi == null no longer means that the HAL service is
no longer running. Add a new boolean to explicitly track this. The flag
is set whenever we retrieve mWifi, but it will be reset only when the
HAL service's death notification is received.

Also, removed some redundant mocking of isReady() API in aware unit tests.

Bug: 124756389
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: Manually verified toggling wifi/softap on/off multiple times.
Change-Id: I80b707f4d3e31364a511b56d76ea6df05b984c53