Use a helper function to get the Ip Memory Store.
This is necessary because the Ip Memory Store will be moving to
the network stack and the way to get one will change from
getSystemService() to just new IpMemoryStore(). The former method
can't be used after the change, but the latter can't be used
before, so we need a method the implementation of which can
change in AOSP for this to merge cleanly.

Context : this should go in after aosp/924884 but before aosp/906907.
After aosp/906907 has gone in we can remove the method and replace
it by a new IpMemoryStore(). Or not.

Bug: None
Test: FrameworkWifiTests
Change-Id: Ia83c23969882b0758badd84d5be1b92ca402490d