Check that network evaluators have reported their choice.
Network evaluators are supposed to call onConnectableListener to report
all networks they want as candidates. This should include the one that
they return.

Add a check in WifiNetworkSelector to ensure this is the case, and
Log.wtf if not.

Add checks to ScoredNetworkEvaluatorTest for all the non-null choices
in the previously existing unit tests. Add a new unit test to cover
the case where a new ephemeral network is created.

Fix ScoredNetworkEvaluator to call onConnectableListener in case of
a new ephemeral network.

Bug: 128938468
Test: atest ScoredNetworkEvaluatorTest
Change-Id: I539dfaa7e5951e19d319a95a90d0b7cc5f3d1a06