Run WifiLock operations from on wifi thread
In current implementatoin the WifiLock operations run on the wifi service
thread pool. Since the application thread blocks on the return of the
that thread, apps go into ANR when the calls block.

This commit runs those operations on the wifi thread. This also removes
the need to use an AsyncChannel for requesting the supported features.

Bug: 129288264
Test: Manual
Test: atest com.android.server.wifi

Change-Id: I08f6e18d532ca14381d0cc1505cbf44c8f3d9608