WifiController: Defer broadcast registration
This should prevent any external interactions being initiated when the
device is in crypt debounce state (converting to FDE). This CL overrides
the start() method in WifiController to perform the broadcast registration
before invoking the parent class (StateMachine) start() method.

Note: WifiController.start() will not be invoked by WifiServiceImpl when the
device is in crypt debounce state.

Bug: 130766132
Test: Device boots & connects to wifi network.
Test: OEM verified the fix.
Test: Regression testing.
Change-Id: Iaa714f5127c7434e57d661643b82b71ed55db6b8