Switch wifi to use IpClientManager.
This is necessary to ensure all calls to IpClient properly clear
the calling identity. Otherwise, when using the in-process
network stack, any operation that originates from another process
(e.g., a request to set the multicast lock) would be denied
because the network stack will only accept requests from the
system UID.

This also simplifies the code by removing the need to use stable
parcelable types and catch RemoteExceptions.

Bug: 131206163
Test: atest FrameworksWifiTests
Test: builds, boots, wifi works
Change-Id: If08f4cdc2b5a4a3d42bfd80aa0cbbe1424e45787