[OWE] Fix Open-OWE transition connectivity for non-supporting devices
Device running Q without OWE support could not connect to Open-OWE
transition network. In InformationElementUti.java We check if OWE
is supported because we are adding the OWE capabilities to the Open
BSS. Non-supporting devices need to see this open network and
ignore this element. Supporting devices need to parse the OWE IE,
hide the Open BSS of OWE in transition, mode and connect to the
Hidden one.

Bug: 132455533
Test: Connect to OWE transition network with non-supporting device
Test: Connect to OWE transition network with supporting device
Test: Connect to OWE network with supporting device
Test: Verify correct scan results on both devices
Change-Id: I2d0d9f68abb36126aae0f011f6bc15a480bcff07