WifiServiceImpl: Use updateInterfaceIpStateInternal for broadcast
updateInterfaceIpState is a privileged WifiManager API that only the
system_server (holding signature NETWORK_STACK permission is allowed to
invoke). We don't need to perform the permission checks when calling the
method for internally handling the AP_STATE_CHANGED broadcast.
Note: updateInterfaceIpStateInternal is already being invoked from other parts
of the handleWifiApStateChanged method.

Note: This is a reland of ag/8107356 without the changes that were causing
regressions.

Bug: 135683539
Test: Manual (changed tethering band selection)
Test: Will send back for regression testing.
Change-Id: I1119cc52cf7446d1e461281bf82601e285ace9e3