WifiScanningService: Use Context.checkPermission
Context.checkPermission ensures that we skip permission checks on
SYSTEM_UID, etc. This is required because the MAINLINE_WIFI_STACK
permission is not granted to any process when we use the
InProcessWifiStack.apk. We still want those permission checks to allow
the wifi stack (running in system_server with SYSTEM_UID) to make those
gated operations.

Bug: 135753701
Test: Verified wifi scanning works (which gates on the
MAINLINE_WIFI_STACK permission).
Test: atest WifiScanningServiceTest
Change-Id: I74efdfb1e1c8e2454fd0364bfd0ae654f756137d