WifiService: Ignore wifi state change API in crypt debounce state
When the device is in crypt debounce state (transitioning to full disk
encryption), we don't start wifi stack. Hence WifiController state
machine is not initialized in this state and any public API calls from
settings or other parts of the system will cause a crash. Handle this
state more gracefully by explicitly failing all the state change API's
in this
state.

Bug: 136799715
Test: atest com.android.server.wifi.WifiServiceImplTest
Change-Id: I36a603c01b1421bfa5e015696902946154ea0989