ActiveModeWarden: made EcmState into an abstract base state
ECM state is now a sub-state of every other state (Scan, Disabled, Client).
Other states are tied to the life time of the corresponding ActiveModeManager.
When the state is exited, the ActiveModeManager is shut down.
However, entering ECM state doesn't necessarily mean that Client state has to
be shut down. Thus, ECM state needs to be a sub-state of Client state.
Same logic applies to scanning state.

Bug: 139157226
Test: Unit tests in a following commit
Test: Will submit for integration testing
Change-Id: I2e708a6efd82a0527d129130e8ca357e8214a7c2