[WiFi] Replace internal connectivity checks permission
A number of connectivity checks that protect system-only methods
check for CONNECTIVITY_INTERNAL, but CONNECTIVITY_INTERNAL is a
signature|privileged permission. We should audit the permissions
checks, and convert checks that protect code that should not be
called outside the system to a signature permission. So replace
the permission to NETWORK_STACK or another proper permission.

Bug: 32963470
Test: atest FrameworksWifiTests
Change-Id: I895439752420d6ba19ff42a8f098700da0d43b80