ActiveModeManager: Add a generic mode manager callback
Add a generic mode manager callback for each mode manager to
indicate the following events:
a) start failure
b) stopped
c) started (this is temporary). Will be removed in a future CL when
ScanOnlyModeManager is removed.

ActiveModeWarden only needs to know when the corresponding mode manager
is no longer needed (either due to failure or if it was stopped). So, it
doesn't need to know any of the other intermediate events,

This also avoids mixing the public broadcast constants for Client mode
& Softap with mode management related callback.

Also, added a dump of state machine for each mode manager.

TODO: Will fix unit tests in a later CL.

Bug: 128585344
Bug: 127624451
Test: Manual tests in progress.
Change-Id: Id93361a4dd2ec1d7dd12701228b839920c8b71cb