ActiveModeWarden: Remove ScanOnlyModeManager
Merge the scan only mode into a state inside ClientModeManager.
ActiveModeWarden will perform an in-place switch of ClientModeManager
from scan only to connect state.

Also, client mode manager no longer waits for interface up to trigger
onStarted() callback to indicate that the mode is fully active. This
does not make sense for client interfaces because the iface is setup by
the HAL/wpa_supplicant as a part of interface setup. Interface up is not
a distinctly different step in the initialization sequence. We still
listen to interface down indication for failures.

TODO: Will fix unit tests in a later CL.
TODO: Measure new wifi toggle on/off times.

Bug: 128585344
Bug: 127624451
Test: Toggle wifi/softap/airplane mode on/off multiple times.
Test: Will send for full regression testing.
Change-Id: Id7e0b421bd5f153ad62afccfd1a8d128a8b66254