WifiScanningService: Support multiple scanner impls for single scan
Create a new helper class to track single scan across multiple impls:
a) Pass the scan request to all of the available impls.
Note: We don't attempt to split the original request based on the impl
capability. This would make the implementation lot more complicated.
The lower layers (wificond/driver) should be able to ignore channels/bands
in the original request that cannot be handled by the corresponding iface.
b) Wait till results are reported from all of the impls.
c) Consolidate the results and pass it to external clients.

As long as at-least 1 of the impls scan succeeded, failures from other
impl's are ignored (since some scan results will be available).

Bug: 140111024
Test: atest com.android.server.wifi.scanner
Change-Id: Ib0aa86a28b9a4576cdee10cffcef1c2af2fb7ecd