WifiScanningService: Add new new impl only for new bands
If a new interface is setup, ensure that a new impl is created only if
the new iface offers a new set of channels/bands.

We want multiple scanner impls to support client intefaces on different
wifi chips which offer distinctly different set of bands to scan. If
multiple client interfaces are created on the same wifi chip, we don't
need multiple impls (since the same scan results can be obtained by the
existing ones).

Bug: 140111024
Test: atest com.android.server.wifi.scanner
Change-Id: Ie2bcf236922f70d3cdabfc2c7ca4df84ee671222