WifiNetworkFactory: Send match callback on registration
If the UI registers its callback to the platform and if we already have
some matching results for the request, send it immediately to the UI.
Otherwise, we'll need to wait for the next scan results to present users
with matches.

Bug: 134712530
Test: atest com.android.server.wifi
Test: Manually verified with CtsVerifier tests that the UI does not
start out with a blank page (because the matched networks are already
found by using the cached results).

Change-Id: Ice7f89952c0a71c181ea5de58e417c02e7f2bf91