Wifi usability: Clear the ring buffer after adding each usability stats to the upload list.
The LABEL_GOOD and LABEL_BAD data may be mixed if the ring buffer is not cleared after each addition of usability stats the upload list. Currently we rely on data post processing to sort them out.  The load on data post processing can be reduced if the ring buffer is cleared.

Bug: 143219266
Test: atest WifiMetricsTest

Change-Id: I24bef0f7d3ce47696e9d84eb5e516c5ee2b7df50
Signed-off-by: Mingguang Xu <mingguangxu@google.com>