Extend the timeout to obtain IpClient
In some configurations (especially after a cold boot) it can take more
than 10 seconds between the time WifiService starts, and NetworkStack
starts. If the phone tries to connect to WiFi in that interval, it is
possible that obtaining IpClient will timeout.

Extend the timeout to one minute to make sure that even in a cold boot
on a slow device, there is enough time for NetworkStack to start.

Test: manual: phone boots, WiFi works
Bug: 142842833
Change-Id: I757de74f54d4d35e67a9abac4746f77a56145fcf