WifiKeystore: Migrate to public keystore
a) Move keystore access to public keystore API surfaces. Some of these
API's are not a 1:1 mapping. For example, the public surface has a
single API for storing the private key and the user certificates.
b) The public keystore API stores the CERT in DER format, so need to convert to
PEM in the wifi keystore HAL.
c) The public keystore API internally appends the keystore alias
prefixes. So, stop appending it in the wifi stack.

Bug: 142089671
Test: Connected to passpoint networks
Test: Will send for full regression tests.
Test: atest com.android.server.wifi
Change-Id: I281f92d7dc4a042c206fb77ae9290a663fafbb6f