WifiDiagnostics: Use shell command to trigger bugreport
a) Move bugreport collection by WifiDiagnostics behind an OEM specific
overlay.
b) Use the shell intent to trigger the bugreport instead of using the @hide
API in ActivityManagerService.

Note: This is not using a formal API, discussions ongoing on whether we
can use a formal API instead in the linked bug. This unblocks to get
feedback on wifi issues from early dogfooders in R.

Bug: 143494985
Test: atest com.android.server.wifi.WifiDiagnosticsTest
Test: Verified bugreport taking using the new mechanism.
Change-Id: I4e44d61d33e456fd0252dba7e382e0feb87de544