wifi-service: Convert to java_library
a) Move the resources to an APK which is linked by the wifi-service java
library.
b) Load the resource values lazily since the resources APK will not be
available when the wifi service jar is initialized.

Also, unit tests now compile the service sources directly instead of
linking wifi-service library. This is because the wifi-service library
linked is jar-jar'ed now which causes compilation failure when
attempting to compile the vanilla test sources.

Bug: 144722612
Test: Device boots up and connects to wifi networks.
Test: atest com.android.server.wifi

Change-Id: I1c20c41b9e6669fbbae0ff9d6f70e0ec30f0d073