WifiPermissionsUtil: Handle location mode check failures
Wifi service is now again starting very early in the bootup
sequence. Location service may not be ready to respond to it's public
API calls from wifi during it's initialization sequence (seen in
wifi-service & rtt-service initialization sequence). Handle any errors
seen gracefully and fallback to the legacy Settings value to read the
location mode setting of the device. The legacy settings value only
reflects the location mode of the user 0, but this is ok early on in
the bootup sequence since any user switches will happen after this.

Bug: 144722612
Test: atest com.android.wifi
Change-Id: I5df9cdffd14962847a8ab23ed54510fafa13a149