Wifi: Replace dual_band with 5GHz band support
As part of the support of the 6GHz band, the concept of dual-band
is no longer valid since we have more than 2 bands supported by
Android and can be checked on the device.
This commit replaces checks on dual-band support with an explicit
check on 5GHz band support.
This commit also reads the device support for 5GHz/6GHz bands through
the overlay configs, and removes the logic for reading it from the
vendor HAL.

Bug: 145936758
Bug: 139354972
Test: Manual
Change-Id: I9f0ad7c5ab57107c71b13f857994e874a34d9bdf