wifi: Remove libwifijni & use logcat binary instead
The only method used in the jni library is a call to fetch kernel logs
when a disconnect occurs. We should be able to run the logcat binary
with appropriate command line flags to get that info instead.

Changes in the CL
a) Remove libwifijni library & its dependencies.
b) Remove WifiNative.readKernelLog() & instead to use the existing logcat
runner with "-b kernel".

Bug: 145196311
Test: Device boots up and connects to wifi networks.
Test: Verified that the kernel logs are present in the bugreport on
every disconnect in the wifi diagnostics section
(modified the code to trigger a WifiDiagnostics report on every disconnect)

Change-Id: Ia2d5c8f6fc509015dd0dbb174069a2c24fa0117e