Change default restriction on 2GHz SAP channels
Current codebase has the configuration default to limit SoftAP
channels in 2GHz band to 1,6,11.
This commit changes this restriction to channels 1-11.
OEMs can override this default by setting their own restriction
in the overlay file.
This allows ACS algorithm in driver to use more channels for SoftAp
in cases of coex issues.

Bug: 147486468
Test: Manual: Start SAP and verify hostapd.conf file to have channels 1-11
Test: atest com.android.server.wifi
Change-Id: I233dfe2c1d011a8b230c744b853c71bd995e2b26