[WifiCandidates] Remove the nominator score
The nominators are no longer generating scores, so do not bother to
store them.

Relax the candidate replacement rules so that new candidate can replace
one with a numerically lower (or equal) nominatorId. This is to allow
a candidate for the current connection to be added as a default, with
the possibility of replacing it using via the nominators.

Bug: 147751334
Test: atest WifiCandidatesTest WifiNetworkSelectorTest
Change-Id: I9d231ecccc6821e8446b4047fdd7509691859ecd