wifi.rc: Package wifi.rc separately into the build
These contain a bunch of action triggers for setting some kernel flags
for wifi driver logging. Apex'es don't support .rc with action triggers,
so package it separately into the build (it does mean that this .rc is not
updateable via wifi apex, but don't see a pressing need for it).

Bug: 149909892
Test: Device boots up and connects to wifi networks.
Test: Manually verified that the .rc file is run on bootup.
Change-Id: Idf4a825170c68899bdfc3712958b3299a73d6672