WifiSettingsConfigStore: Rework to add type safety
Reowrked this class to ensure that the type of value for each key is obvious to
the users of the class. Added a new base class called Key and added
type specifc Key child classes with the constructor marked private. All
the existing keys are declared in these corresponding type specific child
classes. Since the constructors are marked private, no one externally
can create new instances of this class and they have to use the publicly
declared key instances.

Bug: 149738301
Test: Device boots up and connects to wifi networks.
Test: atest com.android.server.wifi
Test: Manually toggled wifi verbose logging/scan throttle flags and
ensured that it's persisted correctly after reboot.

Change-Id: I210d753801ffb1dc46b98aa0aa782209612a4617