Wifi: remove PMK cache in supplicant due to sim removal
When successfully connecting to an EAP-SIM based wifi network, BSSIDs
will be added to PMKSA cache for faster (re-)association. If the SIM
card is removed, wifi connection is brought down.
However, as supplicant still keeps the BSSIDs in its PMKSA cache, a
manual attempt to reconnect to the EAP-SIM network will succeed
even though the SIM card is not present. To disallow this, clear PMKSA
cache in supplicant as well when SIM is removed.

Bug: 149272637
Test: atest FrameworksWifiTests
Change-Id: I7f3ac06d9fdc948b451d3abc17f03802c6529bea