ConnectedNetworkScorer: Do not send invalid sessionId
Generate a sessionId for the external scorer that is not the same as
the android.net.Network netId, to avoid any assumptions that they are
the same.

Assure that the external scorer is never sent an invalid sessionId,
and that a previous session is stopped before a new one is started.

Bug: 122133502
Test: atest android.net.wifi.cts.ConnectedNetworkScorerTest
Test: atest com.android.server.wifi.WifiScoreReportTest
Change-Id: I008d26490df10e3037e1c99025f72c21e3e335d9