Mitigation of incorrect maxNumSpatialStreamDevice.
Many android devices will set maxNumSpatialStreamDevice incorrectly to 8
due to a WLAN driver bug. The bug fix is being ported but not expected
to be applied to all affected products on time. As a mitigation method,
add overriding enable flag and value in config overlay file to override
maxNumSpatialStreamDevice to a correct value before the bug fix is
completely populated. The enable flag is set by default with a default
overriding value of 2. Final settings can be configured per device.

Bug: 148982542
Test: atest com.android.server.wifi
Change-Id: I7b95a0b7f5685ab8b841113ff7136723d8b04250