[Suggestion] Ignore open saved network
When saved open network has a same network suggestion which is from an app with
NETWORK_CARRIER_PROVISIONING permission, also that app suggested a
secure network suggestion with same carrierId and a higher or equal
priority is available to connect. The saved open network will be ignored
during the network selection.

Bug: 154654080
Test: atest com.android.server.wifi

Change-Id: If170fd4ef23d049453b28e655b8305b95307fa01