wifi: Indicate AUTH_FAILURE for WEP password error
For WEP password error, supplicant reports ASSOC_REJECT, this
will not indicate password error to GUI and instead bssid is added
to blacklist on continous failures.

Similar to WPA3, check the status code in Assoc reject event
and if it is "Authentication rejected because of challenge failure",
indicate AUTH_FAILURE for WEP password error.

Bug: 153923670
Test: Manual - Try Connecting to wpa3 and wep network with wrong password
Test: atest com.android.server.wifi
CRs-Fixed: 2433524
Change-Id: Ibe0902ee832504013c2017ecc2442f4580b847e3