WifiNetworkFactory: Don't use WifiConfig from UI for matching
The WifiConfiguration passed in from the UI in the
NetworkRequestUserSelectionCallback.select() only has the SSID field
set (in hindsight the select() API should just have accepted a string
param instead of WifiConfiguration). The rest of params including
security params are constructed from the original app's request. Ensure
that we use the merged WifiConfiguration to find matching scan results
to set bssid for connection.

Also, fixed the unit tests to reflect the correct behavior from UI (only
set the SSID field in WifiConfiguration). The unit tests were
incorrectly setting the securiy params and masking the problem.

Bug: 154764757
Test: act.py -c wifi_manager_cross.config -tb dut-name -tc WifiNetworkRequestTest
Test: atest com.android.server.wifi
Change-Id: I611e05f15a0e1fff3cdbc1f6d0e2347653446678