SupplicantStaIfaceHalCallback: Add ssid in disconnect/assoc-reject events
The current onDisconnected()/onAssociationRejected() callback does not
contain the SSID of the network for which the event was received. This
information is however available in the onStateChanged() callback. Use
the start of the association to capture the current ssid. This is in
lieu of adding a new HAL callback (for ex: a new onDisconnect cb with
ssid, etc) since that would only be available on new HAL versions. The
ssid info in the disconnect/assoc-reject events will help us
disambiguate disconnect events when switching connection between
networks.

Also,
a) Removed unused roaming attempt timestamps in ClientModeImpl.
b) Cleaned up StateChangeResult fields with annotations.

(cherry-picked from ag/11396038)

Bug: 135503681
Test: atest com.android.server.wifi
Test: Manual connect/disconnect, examine logs to ensure that info is
captured.

Change-Id: I68f06bd9a9a05571629b297761982230101d1c7c