Avoid changing wifi mac address when device does not support randomization
Devices without a vendor hal do not provide a way for the wifi framework
to get the factory mac address, so switching from a random mac to
the configured mac requires a reboot. So if mac randomization has not
been enabled in the device config, don't attempt to change the mac.

Bug: 146507767
Test: atest ClientModeImplTest
Change-Id: I53212dc05813001d0ccabeeb4cd491dec06929a4