wifi: Register for IMS network callbacks to delay disconnection.
Wifi disconnect is required to be delayed when Wifi Calling is in
progress and user turns off wifi. ClientModeManager registers for
registerImsRegistrationCallback() but onUnregistered() callback is
received even before IMS PDN is tear down.

This commit is to monitor for IMS NetworkCallbacks and disconnects
wifi only when onLost() indication is received.

On receiving onLost() indication, add additional delay based on
overlay option 'config_wifiDelayDisconnectOnImsLostMs' (default 0).
This is to insure IMS PDN on modem is cleaned up.

Additionally, check all Subscription IDs for active VoWifi call in
progress.

Bug: 143612699
Test: atest FrameworksWifiTests
Signed-off-by: Purushottam Kushwaha <quic_pkushwah@quicinc.com>
Change-Id: I6f2e6271e52a39b4f5115fc84727153e240f5a3b