ClientModeImpl: Remove DisconnectingState
There is an existing timeout (which is still in DisconnectingState)
which was being used to ensure that a new connection attempt times out
if supplicant ignores the request. Move that timeout to the
ConnectingState now since that is the new transient state. Increase the
timeout to 30s to avoid prematurely aborting connection attempts
(EAP connections might take a long time).

Also, remove bunch of unused fields in ClientModeImpl.

Bug: 135503681
Test: atest com.android.server.wifi
Test: Presubmit ACTS tests
Change-Id: Ib331585b73fa764e85e97cae6e16495cccbaedca