WifiNetworkFactory: Use ActiveModeWarden to request ClientModeManager
This will let network factory handle peer to peer connections on either
the primary STA iface or a dedicated local only STA iface (if the device
supports STA + STA when the request comes in).

Note:
a) We can only use the 2nd STA iface once we are able to create 2
instances of ClientModeImpl. This is still being worked on.
b) If we have overlapping requests, use the cached ClientModeManager
instance since the cost of discarding the old one & creating a new one
might be too high.

Bug: 148935691
Test: atest com.android.server.wifi
Test: act.py -c wifi_manager_cross.config -tb dut-name -tc
WifiNetworkRequestTest

Change-Id: I355bbda3ee0f68964241271e59936cc5605d1277