ConfigurationMap: Don't cache specifier ScanResultMatchInfo
WifiConfiguration's created from WifiNetworkSpecifier requests should
not be cached in the internal ScanResultMatchInfo cache maintained by
WifiConfigManager. This WifiConfiguration is truly ephemeral in nature
and is deleted after every connnection. Also, WifiNetworkSpecifier
WifiConfiguration is the only case today where we can have multiple
WifiConfiguration for the same network (because we
WifiConfiguration.shared = false for WifiNetworkSpecifier networks)

Note: A more generic fix for this will be done in b/142035508.
Since not storing cache entries for suggestions or WFA networks may
cause other side effects and too risky for R at this stage.

Bug: 157609789
Test: atest com.android.server.wifi
Test: atest ConnectedNetworkScorerTest WifiNetworkSpecifierTest
Test: act.py -c wifi_manager_cross.config -tb dut-name -tc
WifiNetworkRequestTest

Change-Id: I85cc20bd910cf3e974746653737a8fda54a27e09