WifiDiagnostics: Remove blocking call to readLine()
Changes in the CL:
a) Add a check for isReady() before invoking blocking readLine().
b) Add a timeout in the loop to ensure we're not looping forever.

Note: Even with this, there is a theoretical chance that we will get
stuck in readLine() since isReady() will indicate some bytes available,
but not necessarily EOL. But, the logcat daemon outputs a line of data
at a time.

Bug: 157062041
Test: Manual verification of wifi dumpsys after a connection failure.
Change-Id: I6bbdf5f2fe7ed6afc9dd65d95fcb9e938366fdad