ActiveModeWarden: Defer APM toggle if handling a previous toggle
Back to back mode toggles (wifi, softap) for regular modes are handled by
the respective ActiveModeManager (For ex: ClientModeManager handles the
wifi toggle debounce while IMS dereg is ongoing). However, for back to
back airplane mode toggle handling, the ActiveModeWarden needs to know
if there is an ongoing stop to defer processing of toggle on.

Bug: 157711806
Test: atest com.android.server.wifi
Test: Sent the patch to OEM for verification.
Change-Id: I9595be86d823c7afc5b019efab8d5ffbb834a90b