Let FrameworksWifiTests to use services.net
This is considered as a partial revert of
I11584d3bc87b20f1861bf1a3c3da30b45a250535.

The test was using services.net-module-wifi which in turn depends on
netd_aidl_interface-V3-java. This is actually causing problem because
another version of the AIDL interface netd_aidl_interface-unstable-java
was statically linked to it via the follwing dependency chain:
FrameworksWifiTests
-> services.core
-> services.core.priorityboosted
-> services.core.unboosted
-> netd_aidl_interfaces-platform-java
-> netd_aidl_interface-unstable-java

According to b/157869568, it's not right for a process to have two
different versions of the same AIDL interface stub libraries. Build
system will start to detect and prevent it.

Therefore, this change makes the test to use services.net that uses
netd_aidl_interface-unstable-java.

Bug: 157869568
Bug: 145825329
Test: m

Change-Id: I6e00bf7bb925a38ba1598b4d3c283f7bba8c37df