FrameworksWifiTests: statically link service-wifi
FrameworksWifiTests currently includes all the source
files from service-wifi. Instead, just statically link
service-wifi. This has a couple of advantages:
1. No need to maintain the same list of dependencies in two
   places (service-wifi & FrameworksWifiTests).
2. if service-wifi is already built, no need to recompile
   the same source files again for FrameworksWifiTests.

To accomplish this, remove the dependency on services.core
and instead copy the single method that we need.
Including services.core pulls in a conflicting dependency to
netd_aidl_interface that is a different version from the copy
of netd_aidl_interface pulled in by service-wifi.

Bug: 144722612
Bug: 157869568
Test: atest FrameworksWifiTests
Change-Id: I9e1bc83fbfefb8dc1952eb660ce8b746bb6ab0ce