PasspointManager: Remove usage of ClientModeImpl
The only usage of ClientModeImpl in the class is to trigger a disconnect
if we're connected to a profile corresponding to the app removed. This
explicit trigger is not needed since |removeProvider| already internally
removes any active WifiConfiguration from WifiConfigManager. This in
turn triggers the OnNetworkUpdateListener.onNetworkRemoved() callback
which is used by ClientModeImpl to trigger disconnect.

Bug: 159052883
Test: atest com.android.server.wifi
Change-Id: Iabd827ff0309543127e8b7ef16932174195f07bd