ClientModeImpl: Remove async channel with WifiServiceImpl
The purpose of the aync channel was just to send blocking messages (for
responding to API calls) to the ClientModeImpl's state machine. Most of
these messages (except syncGetCurrentNetwork()) are handled when wifi is
on, else failed. With the new architecture, when wifi is off, these public
API calls will no longer be plumbed to ClientModeImpl, so we can get rid
of this AsyncChannel.

Note: All the CL's in this stack will be submitted together. This CL
does not handle these API calls correctly when wifi is off, will be
fixed in followup CL's.

Bug: 159060934
Test: atest com.android.server.wifi.
Change-Id: I98a0f25e948acafba3a3e25046122b65091c5aa2