ActiveModeWarden: Store external scorer for primary CMM
WifiScoreReport class will be instantiated per instance of CM
(ClientModeManager). Hence we need to cache the external scorer in
ActiveModeWarden and apply it to only the primary ClientModeManager.

The purpose of the external scorer API is to control scoring of the
default wifi network on the device. It is not allowed (or wants) to
control other non-primary client mode managers which maybe present for
STA + STA use-cases.

Bug: 159060934
Test: atest com.android.server.wifi
Change-Id: I04857b1af12ca8e7ff4097f528d0f3fda5b6f6bf