resetSimAuthNetworks: move mode-independent logic out of ClientModeImpl
Updating Wifi(Config|Suggestion)Manager|SimRequiredNotifier is mode
independent. Move it out of ClientModeImpl. Note that part of the logic
is mode-dependent, and that part remains in L2ConnectedState.

Note: the previous logic in ConnectableState should have been in
DefaultState to begin with. It was likely placed in ConnectableState
because when it was implemented in Android N, updating networks was
not possible when Wifi was disabled since wpa_supplicant would be
off, and that is where networks were stored back then.

Bug: 160752427
Test: atest FrameworksWifiTests
Change-Id: Ic7a0e600eed43e1f0474e230bde48db6eee92d65