WifiServiceImpl: Pass in WorkSource toggle API's
Plumbs in WorkSource for various wifi toggle/softap toggle API's
to ActiveModeWarden. Will plumb the worksource to ModeManager's for
plumbing it to HalDeviceManager for prioritization.

Note: Some of the paths for wifi toggle assumes user toggled it which
may not be correct (for ex: after a reboot, we have no clue if the wifi
was toggled on by an app or user before reboot). If we need this info,
we will start storing this info in persistent store. Not sure if the
complexity is warranted since the wifi toggle is being slowly deprecated
for 3p apps.

Bug: 162344695
Test: Compiles
Test: Unit test (TBD)
Change-Id: Ifd411b1282ba0cafa5359985be1d4b8c9b129b33