WifiConnectivityManager: Rework connecting or connected check
WCM should not trigger a new connection attempt if we are already
connecting or connected to the network. Rework the current check to make
sure we cover all cases. Also, stop using WifiInfo for figuring out
internal state of ClientModeImpl. This mixing of params from WifiInfo &
ClientModeImpl causes stuff to go out of sync.

Also, fixed the existing roaming unit tests in WCM.

Bug: 162555276
Bug: 160288009
Test: atest com.android.server.wifi
Test: ACTS tests (TBD)
Change-Id: I5a25666f75ab27d8a03ed37f3ead4f206a148265