WifiP2pService: Stop using HDM iface available callback
We are reworking HDM prioritization scheme. The interface available
callback no longer works in this scheme since we need the app's context
to determine if the P2P iface is available or not for them. The
P2P_AVAILABLE callback will be sent to all apps when wifi is enabled.
When an app tries to initialize the P2P stack, the request is
denied/granted by HDM based on which app triggers the request.

Bug: 162344695
Test: atest com.android.server.wifi.p2p
Change-Id: I1119605ba01d9eef4a20e77838780c374cda6a7d