WifiAwareService: Stop using HDM iface available callback
We are reworking HDM prioritization scheme. The interface available
callback no longer works in this scheme since we need the app's context
to determine if the P2P iface is available or not for them. The
AWARE_AVAILABLE callback will be sent to all apps when wifi & location
is enabled.
When an app tries to initialize the aware stack, the request is
denied/granted by HDM based on which app triggers the request.

Bug: 162344695
Test: atest com.android.server.wifi.aware
Change-Id: I55333463d155e6a0d2fc03469575e88c0913f80b