Don't block bssid if its the only available one
For some failure types, we want to skip blocklisting in the
BssidBlocklistMonitor if the failed BSSID is the only BSSID.

Benefits:
1. Fixes the problem that the BSSID gets blocked before the
WifiConfiguration is disabled if there is only 1 BSSID around. This
issue causes the UI to not display the failure reason text.
2. In case there are multiple configuration for the same network, this
will allow other potentially good configurations to be tried.

Bug: 169272385
Test: atest com.android.server.wifi
Change-Id: Idd8b507185fffa78258000d630b515592e54e348