Fix WifiConfiguration#getAuthType
Fix WifiConfiguration#getAuthType and add unit tests. The root
cause for the crash is due to the outdated implementation of this
method. When adding a suggestion of a WPA2/3-Enterprise network
via the the ACTION_WIFI_ADD_NETWORKS action, Settings calls this
method. WPA2/3-Enterprise networks have more than 1 bit set in
the allowedKeyManagement bitset which the old implementation
treated as an error and threw an exception.

Bug: 169626258
Test: atest WifiConfigurationTest
Merged-In: I85796299f119b84781148c949d701d10da506dc2
Change-Id: I00e112388b392868b0aa9e0ccb8a9b72236bfa2c