ClientModeImpl: Send NETWORK_STATE_CHANGE on roam/new provisioning
These were unintentionally removed in a commit to cleanup the
NetworkAgent API surface (as a part of networking team's API refactor
for mainline). The previous behavior was technically correct because
we don't need to send this L3 network change broadcast on roam events
(since most often that does not cause an L3 change). But, this
changes behavior on a public API surface and apps relying on
this old behavior will be broken.

Bug: 170852221
Bug: 172833539
Test: atest com.android.server.wifi
Change-Id: Ia97ef8714ee2c2b79f305d4f32fef7b5b128dfef
Merged-In: Ia97ef8714ee2c2b79f305d4f32fef7b5b128dfef