softap: add support for SoftAP on the 60GHz band
Support starting SoftAP on the 60GHz band by setting
the apBand field in WifiConfiguration to AP_BAND_60G.

Extend the WifiNative.setupInterfaceForSoftApMode with
band argument. When band is AP_BAND_60G, locate a WIFI
chip which supports the 60GHz band by checking for the
new WIGIG chip capability. If a chip was found, create
the AP interface on this chip.

Extend the channel and band selection methods in
ApConfigUtil to support selecting 60GHz channels.

Implement WifiManager is60GHzBandSupported for detecting
whether 60GHz band is supported in the system. Note,
this method uses a configuration parameter and does not
check for capable WIFI chip. This is because the API
may need to be called when WIFI is disabled, before
the WIFI HAL is loaded. This can be switched to querying
the HAL in the future, when it is loaded earlier.

Add support for configurable 802.11ax (WIFI6) setting,
and disable it by default when using the 60GHz band since
it causes issues with hostapd.

Bug: 155421131
Test: Manual
Test: atest com.android.server.wifi
Change-Id: I82d700de0df1632d55169102372c189b72fc763b