ConnectToNetworkNotification: Disallow connect from lock screen
Avoids malicious usage of the notification to add a saved wifi network
without unlocking the device.

On Android R: Hide the notification in lock-screen.
On Android S & above: Mark the connect action as requiring
authentication.

Bug: 172584372
Test: Manual:
On Android R:
ONA notification not shown in the lockscreen. Visible when the device
is unlocked.
On Android S:
ONA notification shown in lockscreen. Connect button click trigger
device unlock and connects after unlock.

Change-Id: I760cd4303b846a214737edc49254f6327e05d357