ClientModeImpl: Use WifiInfo to get current bssid
WifiInfo fields are updated in `handleSupplicantStateChange()`, so use
these fields to update the layer 2 information.

Alternative is to only use mLastbssid in `updateLayer2Information()`,
but that would need changes in where this method is invoked (which seems
more risky). This is because `mLastBssid` is updated in other places
(not in `handleSupplicantStateChange()`).

Bug: 131797393
Test: atest com.android.server.wifi
Change-Id: I5a35b4863d1bd87ed77cf16bb9c24a6d75478ff0
Merged-In: I5a35b4863d1bd87ed77cf16bb9c24a6d75478ff0