DO NOT MERGE: WifiInfoTest: Reverse setTrusted polarity (2/2)
Default value of trusted bit in WifiInfo is false (which is not the
default used elsewhere like WifiConfiguration, since all networks are
considered "trusted" by default). Reverse the set value to test the
equals funtion.

Note: This fix is only needed in AOSP since master uses another field
(subscription Id) for the same test. Couldn't use that bit in AOSP since
it's not present here.

Bug: 162602799
Bug: 176391251
Test: Will trigger forrest run with wifi unit tests
Change-Id: I6e860f173f53f797a90b25d48762423e8d5fb3b6