ClientModeImpl: Add support to roam between different SSIDs
This change enables roaming between different SSIDs on below
 - Both SSIDs AKM is WPA/WPA2-PSK
 - Both SSIDs have same default gateway MAC address

Currently framework is linking SSIDs on above conditions met and there
is an infrastructure to configure such SSIDs as allowlist SSIDs to
firmware to enable roaming between them. But framework is configuring
always empty allowlist SSIDs to firmware.

This change is to enable generating allowlist SSIDs and configuring them
to firmware based on linked SSIDs of current connected SSID. This commit
adds below changes regarding this
- Upon successful network internet validation get connected SSID default
  gateway MAC address  and persist it.
- Link other SSIDs with same default gateway MAC address and configure
  the linked SSIDs configs to supplicant and firmware.

Bug: 172776379
Test: atest FrameworksWifiTests
Change-Id: I0b0a128c09ce796a5f2f61625539ce72e5976047